// src/app/common/sims/types/sims-pack.ts
export type SimsPack = {
  /**
   * The unique code of the pack.
   */
  code: string;

  /**
   * The name of the pack.
   */
  name: string;
  
  /**
   * The icon of the pack.
   */
  icon: string;

  /**
   * The type of the pack.
   */
  type: string;

  /**
   * Whether the pack is enabled.
   */
  enabled: boolean;
}